Unifying tests. expanding format sweep
